render members added to a team from the api
